examples/perf_analyzer/requirements.txt
aiohttp==3.9.0b0
diskcache>=5.6.3
GitPython>=3.1.43
PyGithub>=2.3.0
dash>=2.17.0
pandas>=2.2.0
dash-bootstrap-components>=1.6.0